Create Selection Sort
In this sorting algorithm, we assume that the first element is the minimum element. Then we check if an element lower than the assumed minimum is present in the rest of the array. If there is, we swap the assumed minimum and the actual minimum, else we move on to the next element.